requirements.txt
django
djangorestframework
requests
beautifulsoup4
